+++ new-api/relay/channel/task/doubao/constants.go
@@ -7,7 +7,7 @@ var ModelList = []string{
 	"doubao-seedance-1-5-pro-251215",
 	"doubao-seedance-2-0-260128",
 	"doubao-seedance-2-0-fast-260128",
-	"doubao-seedance-2-5",
+	"doubao-seedance-2-5-260628",
 }
 
 var ChannelName = "ark-doubao"